nn/requirements.txt
streamlit
PyPDF2
langchain
chromadb
google-generativeai
tqdm
faiss-cpu
python-dotenv
